LU-11509 ldlm: scale LFRU sample window from lru_size
lfru_sample_window_size is the number of lock insertions sampled
before LFRU recomputes priv_score_threshold from max_freq.
It controls how the promotion gate adapts after metadata bursts
without filling the priv list with one-touch locks.

Previously the window was LDLM_DEFAULT_LRU_SIZE / 10 (10 * num_cpus),
set only at namespace creation.  On large clients the sample period
was too long to promote valuable locks before they were displaced; on
small clients it was too short and tending to chased transient churn.
It also ignored the lru_size and was not refreshed at runtime.

This patch addresses the issue by periodically recalculating the
window size based on current lru size, setting it to
ns_nr_unused / 10 and clamping it between 32 and 160, the promotion
gate now dynamically adapts to changing workloads.

Use LDLM_LFRU_PRIV_DEMOTE_THRESH (64) to avoid batch priv demotion
oscillation while the cache is still filling.

And a benchmark is performed, refer #62863 for the code.

30/70 hot/cold access, 800/16000 files, lru_size=2400. Both arms
are LFRU; only the sample window changes. The old window is
10*ncpu (640 on 64 cores). The new window is
clamp(nr_unused/10, 32, 160), i.e. based on LRU size and closer
to the workload. Pin lfru_sample_window_size to 640 vs 160.

Small-CPU mixed runs already show LFRU ~6% ahead of LRU.
On large-CPU boxes the ncpu-based window updates too slowly
and LFRU looks like LRU; sizing from LRU size keeps that lead.

16-CPU, 4 procs, 10 rounds: LRU-sized 160 cut enqueue ~4% vs
ncpu-sized 640:

| Test Run | ncpu-640 | Time | lru-160 | Time | Improvement |
| 1        | 113401   | 116s | 107994  | 113s | 4%          |
| 2        | 113880   | 113s | 107608  | 112s | 5%          |
| 3        | 113507   | 112s | 108381  | 111s | 4%          |

Priv under the LRU-sized window settled near the 30% cap
(~700 of 2400).


sanity/124e/124f/124h, set llite.*.enable_statahead_fname=0 with the
intention to disable `stat()` randomness, to reduce flakiness.

Test-Parameters: testlist=sanity env=ONLY=124e,124f,ONLY_REPEAT=100
Test-Parameters: testlist=sanity env=ONLY=124g,ONLY_REPEAT=200
Test-Parameters: clientdistro=el10.1 serverdistro=el10.1 testlist=sanity env=ONLY=124g,ONLY_REPEAT=200

+++ lustre/obdclass/class_obd.c
@@ -1218,7 +1218,7 @@ bool obd_counter_add_test(struct obd_counter_instance *instance,
 	} else {
 		overlap = winsz - (time % winsz);
 		val = instance->oci_hist[0] +
-		      instance->oci_hist[1] * overlap / winsz;
+		      (u64)instance->oci_hist[1] * overlap / winsz;
 	}
 
 	threshold_exceeded = val > max;
